CIRC-524 Pull request fixes
- Include minutes in the OverdueFineInterval which allows to remove
  interval minutes map from the OverdueFineCalculatorService
- Extract OverdueFineInterval class
- Move static internal class CalculationParameters to the bottom of
  OverdueFineCalculatorService
